API/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.45
pydantic[email]
envparse==0.2.0
# psycopg2 # ==2.0.10
asyncpg==0.27.0
alembic==1.9.0
python-jose
python-multipart
torch
pandas
opencv-python
scikit-learn
torchvision
torchmetrics